requirements.txt
gradio
transformers
Pillow
requests
torch
